Revert to APP routes
